resolve merge conflicts, refactoring
