+++ tools/Evm/T8n/JsonTypes/RejectedTx.cs
