+++ tools/Evm/T8n/T8nBlockHashProvider.cs
@@ -1,14 +1,14 @@
 // SPDX-FileCopyrightText: 2024 Demerzel Solutions Limited
 // SPDX-License-Identifier: LGPL-3.0-only
 
-using Evm.t8n.Errors;
+using Evm.T8n.Errors;
 using Nethermind.Core;
 using Nethermind.Core.Crypto;
 using Nethermind.Evm;
 
 namespace Evm.t8n;
 
-public class T8NBlockHashProvider : IBlockhashProvider
+public class T8nBlockHashProvider : IBlockhashProvider
 {
     private readonly Dictionary<long, Hash256?> _blockHashes = new();
     private static readonly int _maxDepth = 256;
@@ -22,8 +22,8 @@ public class T8NBlockHashProvider : IBlockhashProvider
         }
 
         return _blockHashes.GetValueOrDefault(number, null) ??
-               throw new T8NException($"BlockHash for block {number} not provided",
-                   T8NErrorCodes.ErrorMissingBlockhash);
+               throw new T8nException($"BlockHash for block {number} not provided",
+                   T8nErrorCodes.ErrorMissingBlockhash);
     }
 
     public void Insert(Hash256 blockHash, long number)
