+++ tools/Evm/T8n/T8nCommand.cs
@@ -2,6 +2,7 @@
 // SPDX-License-Identifier: LGPL-3.0-only
 
 using System.CommandLine;
+using Evm.t8n;
 
 namespace Evm.T8n;
